+++ lib/client/client.ts
@@ -1,3 +1,4 @@
+import { GatewayPresenceUpdate, Snowflake } from "../../deps.ts";
 import { HTTPClientOptions, RESTClient } from "../api/http/mod.ts";
 import {
   GatewayClient,
@@ -6,8 +7,13 @@ import {
 import { Guild, User } from "../structs/mod.ts";
 import { Cache, PartialKeys, Storable } from "../util/mod.ts";
 
+export interface APIPresence extends GatewayPresenceUpdate {
+  id: bigint;
+}
+
 export interface ClientOptions {
   guilds?: Storable<Guild>;
+  presences: Storable<APIPresence>;
   restOptions?: HTTPClientOptions;
   users?: Storable<User>;
 }
@@ -15,14 +21,16 @@ export interface ClientOptions {
 export class Client {
   gateway;
   guilds;
+  presences;
   rest;
   users;
 
   constructor(token: string, options?: ClientOptions) {
     this.gateway = new GatewayClient(token);
-    this.guilds = options?.guilds ?? new Cache<Guild>(Guild, this);
+    this.guilds = options?.guilds ?? new Cache<Guild>(this, Guild);
+    this.presences = options?.presences ?? new Cache<APIPresence>(this);
     this.rest = new RESTClient(token, options?.restOptions);
-    this.users = options?.users ?? new Cache<User>(User, this);
+    this.users = options?.users ?? new Cache<User>(this, User);
   }
 
   async connect(data: PartialKeys<GatewayClientConnectData, "url">) {