+++ lib/mod.ts
@@ -2,4 +2,3 @@ export * from "./api/mod.ts";
 export * from "./client/mod.ts";
 export * from "./structs/mod.ts";
 export * from "./util/mod.ts";
-export * from "../meta.ts";