+++ lib/structs/channels/mod.ts
@@ -11,4 +11,4 @@ export * from "./storechannel.ts";
 export * from "./textchannel.ts";
 export * from "./voicechannel.ts";
 
-export * from "./fromtype.ts";
+export * from "./util/mod.ts";