+++ lib/structs/mod.ts
@@ -1,4 +1,5 @@
 export * from "./channels/mod.ts";
 export * from "./guild/mod.ts";
+export * from "./message.ts";
 export * from "./struct.ts";
 export * from "./user.ts";