+++ mod.ts
@@ -1,1 +1,2 @@
 export * from "./lib/mod.ts";
+export * from "./meta.ts";