+++ lib/api/interactions/interactionsclient.ts
@@ -1,4 +1,0 @@
-export class InteractionsClient {
-}
-
-export default InteractionsClient;