+++ meta.ts
@@ -2,6 +2,6 @@
 export const
   deno = "1.8.3",
   repo = "https://github.com/Apacheli/Space",
-  version = "0.4.1-alpha",
+  version = "0.4.2-alpha",
   mod = `https://deno.land/x/space@${version}`,
   tag = `${repo}/tree/${version}`;